# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API. You can find out more about the API at https://github.com/PacktPublishing/Building-Distributed-Applications-in-Gin.\n\nSchema: http",
    "title": "Recipes API",
    "contact": {
      "name": "Arthur Grigoryan",
      "url": "https://inveteratecoder.github.io",
      "email": "inveterate.coder@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Recipe"
              }
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Creates a recipe.",
        "operationId": "createRecipe",
        "parameters": [
          {
            "name": "recipe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/recipes/search": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Searches recipes by tag.",
        "operationId": "searchRecipes",
        "parameters": [
          {
            "type": "string",
            "name": "tag",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Recipe"
              }
            }
          }
        }
      }
    },
    "/recipes/{id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "summary": "Updates a recipe by ID.",
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "recipe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Recipe not found"
          }
        }
      },
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "Deletes a recipe by ID.",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found"
          }
        }
      }
    }
  }
}